refactor(oauth): config structarize
