+++ src/index.ts
@@ -4,9 +4,7 @@ import express from "express";
 import createLogger from "./lib/wintson.logger";
 import jwtRoutes from "./routes/jwt.routes";
 import sessionRoutes from "./routes/session.routes";
-import setupGoogleRoutes from "./routes/google.routes";
 import setupSession from "./config/session.config";
-import setupGoogleOath from "./config/google.config";
 import setupOAuth from "./config/oauth.config";
 import oauthRoutes from "./routes/oauth.routes";
 import jwtMiddleware from "./middlewares/jwt.middleware";
@@ -34,12 +32,6 @@ function config(config: Config): Router {
   if (config.session && config.session.enabled) {
     setupSession(router, configurations);
     sessionRoutes(router, configurations);
-  }
-
-  // Set up routes if Google OAuth is enabled
-  if (config.google && config.google.enabled) {
-    setupGoogleOath(config);
-    setupGoogleRoutes(router, config);
   }
 
   // Set up OAuth if enabled - SETUP BEFORE ROUTES
